engine-db-query: Add --log and logfile to module
In the past, every command was logged into /tmp/engine-db-query.log
which was created based on the users executing the SQL command.
However, if a privileged user create a file, no other ordinary
user will be able to access the file. This patch adds the option
to specify the log file via --log in the bin. In the engine-db-query
module we created the param logfile, if not specified, the logger
will use the Null handler.

Change-Id: I19d02f0465beddd191815be98913b3cd24a08086
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1609667
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>